# openapiclient.json
{"openapi":"3.1.0","info":{"title":"SpeedShip Api","description":"\n    Delivery Management System\n    ### Seller\n    - Submit Shipment Effortlessly\n    - Share tracking links with customers\n\n    ### Delivery Agent\n    - Auto Accept Shipments\n    - Track and Update Shipment Status\n    - Email Notifications\n    ","version":"0.1.0"},"paths":{"/shipment/":{"get":{"tags":["Shipment"],"summary":"Get Shipment","operationId":"get_shipment","security":[{"SellerAuth":[]}],"parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipmentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Shipment"],"summary":"Submit Shipment","operationId":"submit_shipment","security":[{"SellerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipmentCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipmentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Shipment"],"summary":"Update Shipment","operationId":"update_shipment","security":[{"PartnerAuth":[]}],"parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipmentUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipmentUpdate"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/shipment/track":{"get":{"tags":["Shipment"],"summary":"Get Tracking","operationId":"get_tracking","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/shipment/tag":{"get":{"tags":["Shipment"],"summary":"Add Tag To Shipment","operationId":"add_tag_to_shipment","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Id"}},{"name":"tag_name","in":"query","required":true,"schema":{"$ref":"#/components/schemas/TagName"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipmentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Shipment"],"summary":"Remove Tag From Shipment","operationId":"remove_tag_from_shipment","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Id"}},{"name":"tag_name","in":"query","required":true,"schema":{"$ref":"#/components/schemas/TagName"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipmentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/shipment/tagged":{"get":{"tags":["Shipment"],"summary":"Get Shipments With Tag","operationId":"get_shipments_with_tag","parameters":[{"name":"tag_name","in":"query","required":true,"schema":{"$ref":"#/components/schemas/TagName"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShipmentRead"},"title":"Response Get Shipments With Tag"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/shipment/cancel":{"get":{"tags":["Shipment"],"summary":"Cancel Shipment","operationId":"cancel_shipment","security":[{"SellerAuth":[]}],"parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipmentRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/shipment/review":{"get":{"tags":["Shipment"],"summary":"Submit Review","operationId":"submit_review","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Shipment"],"summary":"Submit Review","operationId":"submit_review","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_submit_review"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seller/signup":{"post":{"tags":["Seller"],"summary":"Register Seller","operationId":"register_seller","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SellerCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SellerRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seller/verify":{"get":{"tags":["Seller"],"summary":"Verify Seller Email","operationId":"verify_seller_email","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seller/forgot_password":{"get":{"tags":["Seller"],"summary":"Forgot Password","operationId":"forgot_password","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string","format":"email","title":"Email"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seller/reset_password_form":{"get":{"tags":["Seller"],"summary":"Get Reset Password Form","operationId":"get_reset_password_form","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seller/reset_password":{"post":{"tags":["Seller"],"summary":"Reset Password","operationId":"reset_password","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_reset_password"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seller/login":{"post":{"tags":["Seller"],"summary":"Login Seller","operationId":"login_seller","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_seller"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/seller/dashboard":{"get":{"tags":["Seller"],"summary":"Get Dashboard","operationId":"get_dashboard","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Seller"}}}}},"security":[{"SellerAuth":[]}]}},"/seller/logout":{"get":{"tags":["Seller"],"summary":"Logout Seller","operationId":"logout_seller","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"SellerAuth":[]}]}},"/partner/signup":{"post":{"tags":["Partner"],"summary":"Register Delivery Partner","operationId":"register_delivery_partner","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryPartnerCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryPartnerRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/partner/verify":{"get":{"tags":["Partner"],"summary":"Verify Delivery Partner Email","operationId":"verify_delivery_partner_email","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/partner/forgot_password":{"get":{"tags":["Partner"],"summary":"Forgot Password","operationId":"forgot_password","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string","format":"email","title":"Email"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/partner/reset_password_form":{"get":{"tags":["Partner"],"summary":"Get Reset Password Form","operationId":"get_reset_password_form","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/partner/reset_password":{"post":{"tags":["Partner"],"summary":"Reset Password","operationId":"reset_password","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"requestBody":{"required":true,"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_reset_password"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/partner/login":{"post":{"tags":["Partner"],"summary":"Login Delivery Partner","operationId":"login_delivery_partner","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_delivery_partner"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/partner/logout":{"get":{"tags":["Partner"],"summary":"Logout Delivery Partner","operationId":"logout_delivery_partner","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"PartnerAuth":[]}]}},"/partner/":{"post":{"tags":["Partner"],"summary":"Update Delivery Partner","operationId":"update_delivery_partner","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeliveryPartnerUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"PartnerAuth":[]}]}}},"components":{"schemas":{"Body_login_delivery_partner":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_delivery_partner"},"Body_login_seller":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_seller"},"Body_reset_password":{"properties":{"password":{"type":"string","title":"Password"}},"type":"object","required":["password"],"title":"Body_reset_password"},"Body_submit_review":{"properties":{"rating":{"type":"integer","maximum":5.0,"minimum":1.0,"title":"Rating"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"}},"type":"object","required":["rating","comment"],"title":"Body_submit_review"},"DeliveryPartnerCreate":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"servicable_zip_codes":{"items":{"type":"integer"},"type":"array","title":"Servicable Zip Codes"},"max_handling_capacity":{"type":"integer","title":"Max Handling Capacity"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","email","servicable_zip_codes","max_handling_capacity","password"],"title":"DeliveryPartnerCreate"},"DeliveryPartnerRead":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"servicable_zip_codes":{"items":{"type":"integer"},"type":"array","title":"Servicable Zip Codes"},"max_handling_capacity":{"type":"integer","title":"Max Handling Capacity"}},"type":"object","required":["name","email","servicable_zip_codes","max_handling_capacity"],"title":"DeliveryPartnerRead"},"DeliveryPartnerUpdate":{"properties":{"servicable_zip_codes":{"anyOf":[{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Servicable Zip Codes"},"max_handling_capacity":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Handling Capacity"}},"type":"object","required":["servicable_zip_codes"],"title":"DeliveryPartnerUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Seller":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"email_verified":{"type":"boolean","title":"Email Verified","default":false},"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"zip_code":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Zip Code"}},"type":"object","required":["name","email","id","created_at"],"title":"Seller"},"SellerCreate":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["name","email","password"],"title":"SellerCreate"},"SellerRead":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["name","email"],"title":"SellerRead"},"ShipmentCreate":{"properties":{"content":{"type":"string","title":"Content"},"weight":{"type":"number","maximum":25.0,"title":"Weight"},"destination":{"type":"integer","title":"Destination"},"client_contact_email":{"type":"string","format":"email","title":"Client Contact Email"},"client_contact_phone":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Client Contact Phone"}},"type":"object","required":["content","weight","destination","client_contact_email"],"title":"ShipmentCreate"},"ShipmentEvent":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"location":{"type":"integer","title":"Location"},"status":{"$ref":"#/components/schemas/ShipmentStatus"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"shipment_id":{"type":"string","format":"uuid","title":"Shipment Id"}},"type":"object","required":["id","created_at","location","status","shipment_id"],"title":"ShipmentEvent"},"ShipmentRead":{"properties":{"content":{"type":"string","title":"Content"},"weight":{"type":"number","maximum":25.0,"title":"Weight"},"destination":{"type":"integer","title":"Destination"},"id":{"type":"string","format":"uuid","title":"Id"},"timeline":{"items":{"$ref":"#/components/schemas/ShipmentEvent"},"type":"array","title":"Timeline"},"estimated_delivery":{"type":"string","format":"date-time","title":"Estimated Delivery"},"tags":{"items":{"$ref":"#/components/schemas/Tag"},"type":"array","title":"Tags"}},"type":"object","required":["content","weight","destination","id","timeline","estimated_delivery","tags"],"title":"ShipmentRead"},"ShipmentStatus":{"type":"string","enum":["placed","in_transit","out_for_delivery","delivered","cancelled"],"title":"ShipmentStatus"},"ShipmentUpdate":{"properties":{"location":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Location"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"status":{"anyOf":[{"$ref":"#/components/schemas/ShipmentStatus"},{"type":"null"}]},"estimated_delivery":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Estimated Delivery"}},"type":"object","title":"ShipmentUpdate"},"Tag":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"name":{"$ref":"#/components/schemas/TagName"},"instruction":{"type":"string","title":"Instruction"}},"type":"object","required":["id","created_at","name","instruction"],"title":"Tag"},"TagName":{"type":"string","enum":["express","standard","fragile","heavy","international","domestic","temperature_controlled","gift","return","documents"],"title":"TagName"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"SellerAuth":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/seller/login"}}},"PartnerAuth":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/partner/login"}}}}},"tags":[{"name":"Shipment","description":"Operations related to Shipments"},{"name":"Seller","description":"Operations related to Seller"},{"name":"Partner","description":"Operations related to Delivery Partner"}]}